create handler module
